requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
Pillow
